A-Bループ機能の装着！

+++ lib/domain/models/video_models.dart
@@ -34,7 +34,9 @@ class VideoInfo {
   double get frameDuration => frameRate > 0 ? (1.0 / frameRate) : (1.0 / 30.0);
 
   double calculateTargetPosition(double currentPosSecs, int frames) {
-    return (currentPosSecs + frames * frameDuration).clamp(0.0, durationSecs);
+    return (currentPosSecs + frames * frameDuration)
+        .clamp(0.0, durationSecs)
+        .toDouble();
   }
 }
